Some edit

+++ src/Models/DataTypes/Vector2.java
@@ -1,10 +1,10 @@
 package Models.DataTypes;
 
 public class Vector2 {
-    public float x;
-    public float y;
+    public double x;
+    public double y;
 
-    public Vector2(int x, int y){
+    public Vector2(double x, double y){
         this.x=x;
         this.y=y;
     }
